[007] rename directory
