+++ jsconfig.json
@@ -7,7 +7,7 @@
 			"lib/*": ["./src/lib/*"],
 			"model/*": ["./src/model/*"],
 			"store/*": ["./src/store/*"],
-			"modules/*": ["./src/modules/*"],
+			"view/*": ["./src/view/*"],
 			"style/*": ["./src/style/*"]
 		}
 	}
